Techdebt: Reenable EC2 VPC endpoint services tests
